Enabled [<Property(QuietOnSuccess = true)>]
in the xUnit.net Glue Library.

The purpose of this is the following: TestDriven.NET is a Visual Studio
plug-in for running unit tests. It runs tests quietly on command, but pops
up the Output Window if any test fails. It also pops up the Output Window
if anything is written to Standard Out. This can be confusing when
accustomed to the appearance of the Output Window indicating a test
failure (it's almost a Pavlowian reaction). Thus, the QuietOnSuccess
property offers an option to tell FsCheck to be quiet if no test failures
occurred.

This feature isn't covered by unit tests, because it looks like the
[<Property>] attribute isn't covered by automated tests, but I've tried
to set the property to true in a manual test and verified that there's no
output from that particular test.