Duplicated the expression that writes to std out
upon test completion, in the NUnit Glue Library. While this introduces
seemingly redundant duplication, this prepares the code for the next
change, which is to make the output upon success configurable.

Since there are no automated tests covering the NUnit Glue Libraries, I
find it most prudent to proceed in baby steps.