Added a QuietOnSuccess property to NUnit Property
attribute.

Although no automated tests cover this, I also added an example of using
this property to the example library, and executed this small test suite
with the NUnit GUI runner, as well as the NUnit command-line runner, and
manually verified that no output was produced by the new noOutputOnSuccess
example Property.