Added special case 4 exception thrown during test.
xUnit.net has particular support for dealing with exceptions during test
(which is the normal way to signal a test failure: outside of FsCheck, the
most common was to fail an assertion in .NET is to throw an exception).

This special case pulls the exception out of the FsCheck result and passes
it to xUnit.net via the AssertException constructor overload.

This means that tooling can use the exception information, including the
stack trace, to go to the offending line of code, from a list of test
failures.