Added information about versioning to README.md.
When troubleshooting versioning problems (e.g. with NuGet), it can be
useful for users to know that they can rely on a particular library to
follow Semantic Versioning. This enables users to assume that there are no
breaking changes among versions sharing the same major version, but that
there ARE breaking changes between versions with two different major
versions.