Removed print-to-console expression.
The single `printf` expression made me curious, because I couldn't find
any other `printf` expressions in the PropertyTestCase class. If this was
the only `printf` expression, then how does all the other output get
printed to the console, e.g. when tests fail?

As far as I can tell, the output to print to the console is passed as
strings to the TestPassed and TestFailed objects, which the test runner
then subsequently uses to print the test results to the console.

In order to verify that theory, I temporarily changed the implementation
of the TestResult.True case to this:

let output =
    if not quietOnSuccess then
        let msg = Runner.onFinishedToString "" xunitRunner.Result
        printf "printed: %s%s" Environment.NewLine msg
        "returned value: " + msg
    else ""

Then I built the entire FsCheck code base, and copied the binaries to a
test project in order to see the output. Running a single test with
QuietOnSuccess = false yields this output:

------ Test started: Assembly: ConwaysGameOfLife04.dll ------

Output from Ploeh.Katas.GameOfLifeProperties.Any dead cell with exactly 3
live neighbors becomes a live cell:
  returned value: Ok, passed 100 tests.

1 passed, 0 failed, 0 skipped, took 1,02 seconds (xUnit.net 2.1.0 build
  3179).

The "returned value: " prefix is included in the message, whereas the
"printed: " prefix is nowhere to be seen. This seems to indicate that the
test runner supresses messages written directly to Standard Out. In any
case, the desired message is still printed as appropriate.

Therefore, I consider it correct to remove the `printf` expression.