Replaced an xUnit.net assertion with Unquote.
Notice that I updated the convolutedGenNumber function to use `n`. Before,
the function took the `n` argument, but ignored it, and instead used the
hard-code list expression [1..100]. Since 100 is also used in the exercise
of the function, I interpreted the meaning of `n` to be the upper limit of
the list.

The reason I didn't leave it as is was that when the `n` argument wasn't
used, the convolutedGenNumber was generic, and quoted expressions can't be
generic.

For this refactoring, I considere piping into `(=!) true`, but by using
the full `test <@ ... @>` expression, Unquote will be much better able to
progressively report on what went wrong if the test fails.