Separated the three phases of a test with a blank line.
Note that, in this case, there isn't an Arrange Phase, since there really
isn't a SUT per se.

The Exercise Phase ends right after we get the `typeClass` value, so
there's only one blank line, between the Exercise Phase and the
Verification Phase.

"Like most other code, unit tests are read more than they are written, so
it's important to make the tests readable.

"The technique is simple:

"As long as there are less than three lines of code in a test, they appear
without any special formatting or white space.

"When a test contains more than three lines of code, you separate the three
phases with a blank line.

"When a single phase contains so many lines of code that you'll need to
divide it into subsections to make it readable, you should explicitly mark
the beginning of each phase with a code comment." -- http://blog.ploeh.dk/2013/06/24/a-heuristic-for-formatting-code-according-to-the-aaa-pattern/

(Quoted as multi-paragraph, as per: http://english.stackexchange.com/a/2314)