Separated the three phases of a test with a blank line.
"Like most other code, unit tests are read more than they are written, so
it's important to make the tests readable.

"The technique is simple:

"As long as there are less than three lines of code in a test, they appear
without any special formatting or white space.

"When a test contains more than three lines of code, you separate the three
phases with a blank line.

"When a single phase contains so many lines of code that you'll need to
divide it into subsections to make it readable, you should explicitly mark
the beginning of each phase with a code comment." -- http://blog.ploeh.dk/2013/06/24/a-heuristic-for-formatting-code-according-to-the-aaa-pattern/

(Quoted as multi-paragraph, as per: http://english.stackexchange.com/a/2314)