Add dependency tracking to OperationResult.
Playing with testing a "process registry" and finding all kinds of
interesting gotchas for testing mutable things. This commit adds the
basic functionality to use an "operation result", really a variable, and
makes sure that shrinking takes the implicit dependencies into account.
That is, if an operation A returns a result that is then used in
operations B (e.g. A spawns a new process and B used the pid) then this
dependency is now recorded and taken into account during shrinking.