Added a property about Gen.shuffle.
With the previous test suite, I could use the Devil's Advocate review
technique to reduce Gen.shuffle to this:

```F#
let shuffle (xs : 'a seq) = gen { return Array.empty<'a> }
```

Despite this degenerate, and obviously incorrect, implementation, all
tests passed. This demonstrated the need for more tests, so I added one.

I'm not 100% certain that the term 'isomorphic' entirely covers this
property, but if we regard the domain as an ordered collection of a
certain size, and the codomain as equal to the domain, then shuffle is
isomorphic because every element in the input is mapped to a position in
the output, and a reverse map exists.