Changed Gen.listOf to stay below size in length.
The previous version added 1 to the size in order to "decrease [the]
chance of empty list". This had the peculiar effect that you'd see lists
longer than `size`:

> [0..9] |> Gen.elements |> Gen.listOf |> Gen.sample 1 9;;
val it : int list list = [[4]; [9]; []; [7]; [0; 6]; [5]; [8]; [3]; []]

Notice the `[0; 6]` sample, which has a length of 2.

According to the documentation, this isn't wrong per se, since it only
states that the "maximum length depends on the size parameter". Still, I
found this unintuitive, and that it would make it harder to explain to
beginners how FsCheck works.

For sizes from 1 to 10, the old implementation had these frequencies of
empty lists:

Size  Empty lists
   1         42 %
   2         37 %
   3         32 %
   4         29 %
   5         26 %
   6         25 %
   7         23 %
   8         22 %
   9         21 %
  10         19 %

In the new implementation, I've fiddled a bit with the numbers to
approximate something similar:

Size  Empty lists
   1         45 %
   2         37 %
   3         32 %
   4         29 %
   5         26 %
   6         24 %
   7         22 %
   8         21 %
   9         20 %
  10         18 %

As one can see, the new implementation creates empty lists with a slightly
higher frequency for small sizes, but with lower frequencies as the size
increases.