Eliminate tuple allocations in splittable-RNG functions.
These changes move the Rnd struct out into the FsCheck namespace and
moved the mixing functions from the Random module into members of the
Rnd struct. The Random module was converted into a static class and the
functions converted to static methods. These methods have been modified
to use 'out' parameters to return the updated Rnd value instead of
returning it with the generated random value in a tuple, thus
eliminating a large source of temporary allocations when running
FsCheck tests (which caused slowdowns due to GC).