Renamed DontShrink to DoNotShrink,
since "dont" isn't an English word; "don't" is a proper contraction of "do
not", but the apostrophe is required. Omitting it is necessary in code,
but "Dont" looks wrong.

The .NET framework design guidelines is also clear on the subject:

"DO NOT use abbreviations or contractions as part of identifier names."

Source: https://msdn.microsoft.com/en-us/library/ms229045