Handled char-based flags enums.
As far as I can tell, char-based flags enums don't make much sense, since
they don't seem to support binary 'or' or 'and'. (If there's a way to 'or'
two char-based flags together, I haven't found out how to do it.) As the
added test demonstrates, however, it's possible to define and compile a
char-based flags enum, so FsCheck ought to explicitly handle such a type.