Corrected an error with conditional compilation (#320)
As far as I can tell, I made an error in my last commit. The intent was to
make HostName available on both the full framework and PCL, but I think I
got it wrong. I hope this commit fixes it.

The call to ToCharArray() is necessary because, apparently, System.String
doesn't implement `char seq` on all portable platforms.