Added PowerPoint source for logos
While it'd have been more professional to create the logos in some
'real' graphics design software, I did it in PowerPoint, because that's
what I'm familiar with.

The chosen logo is slide number 3, but the second slide you'll see if
you run the slide show, since slide number 2 is hidden.

I've kept all the slides I used for the various suggestions posted in
https://github.com/fscheck/FsCheck/issues/219 in case anyone needs them
for some purpose in the future.

When creating the logo files, I open the PowerPoint file, select the
figure(s) in question, right-click and choose 'Save as Picture'.

Note that most of the figures are surrounded by an invisible square. Be
sure to include that square when saving as picture, because if you
don't, the exported picture will not be quadratic.